Switching to librewolf, to newsraft and modifying MonitorRule's rate from float to double bc compiler screams at me and i am scared

+++ dwl/dwl.c
@@ -249,7 +249,7 @@ typedef struct {
 	int x, y;
 	int resx;
 	int resy;
-	float rate;
+	double rate;
 	int mode;
 	int adaptive;
 } MonitorRule;
